get_field() corrections
Now properly returns a single field, regardless of how many were specified.

If a single field is specified, then its values are returned in a simple index array.

If more than one field is specified, the first is used as the index to an assoc. array, and the second is the "target" (ie, the basis of the value).

get_field() now passes the "target" field to reindex_by_pk() (whereas get() does not), whereby reindex_by_pk() knows better what to return and in what form.